chore!: Update Package structure
